requirements.txt
asgiref==3.10.0
channels==4.3.1
channels_redis==4.3.0
Django==5.2.7
djangorestframework==3.16.1
msgpack==1.1.2
mysqlclient==2.2.7
python-decouple==3.8
redis==6.4.0
sqlparse==0.5.3
